Borrando Migracion

+++ Migrations/20200407000022_Inicial.cs
@@ -147,7 +147,7 @@ namespace SistemaVentas.Migrations
                         .Annotation("Sqlite:Autoincrement", true),
                     CobroId = table.Column<int>(nullable: false),
                     VentaId = table.Column<int>(nullable: false),
-                    Monto = table.Column<float>(nullable: false),
+                    Monto = table.Column<decimal>(nullable: false),
                     CobrosId = table.Column<int>(nullable: true)
                 },
                 constraints: table =>
